spykes/payment/unversioned-requirements.txt
Django
djangorestframework
stripe